Add gitignore and blacklisting to file deletion

+++ packages/cli/src/utils/files.js
@@ -6,6 +6,7 @@ const path = require('path');
 const fs = require('fs');
 const fse = require('fs-extra');
 const colors = require('colors/safe');
+const constants = require('../constants');
 
 const fixHome = (dir) => {
   const home = process.env.HOME || process.env.USERPROFILE;
@@ -201,7 +202,7 @@ const makeTempDir = () => {
   return workdir;
 };
 
-async function deleteUnmatchedFiles(src, dest) {
+async function deleteUnmatchedFiles(src, dest, ig) {
   try {
     const [filesInTarget, filesInSource] = await Promise.all([
       fs.promises.readdir(dest, { withFileTypes: true }),
@@ -214,12 +215,18 @@ async function deleteUnmatchedFiles(src, dest) {
       const targetPath = path.join(dest, file.name);
       const sourcePath = path.join(src, file.name);
 
+      // if the file is gitignored or blacklisted, do not delete it
+      const relativePath = path.relative(dest, targetPath);
+      if (ig.ignores(relativePath) || file.name === '.gitignore' || constants.BLACKLISTED_PATHS.includes(file.name)) {
+        continue;
+      }
+
       if (file.isDirectory()) {
         if (!sourceFiles.has(file.name)) {
           await fs.promises.rm(targetPath, { recursive: true, force: true });
           console.log(`Deleted directory: ${targetPath}`);
         } else {
-          await deleteUnmatchedFiles(sourcePath, targetPath);
+          await deleteUnmatchedFiles(sourcePath, targetPath, ig);
         }
       } else {
         if (!sourceFiles.has(file.name)) {
